+++ src/app/features/products/components/product-list/product-list.component.ts
@@ -4,6 +4,8 @@ import { take } from 'rxjs';
 
 import { ProductModel } from 'app/features/products/models/product.model';
 import { ProductsService } from '../../services/products.service';
+import { Router } from '@angular/router';
+import { CartService } from 'app/features/cart/services/cart.service';
 
 @Component({
     selector: 'app-product-list',
@@ -13,10 +15,22 @@ import { ProductsService } from '../../services/products.service';
 export class ProductListComponent implements OnInit {
     productList: ProductModel[] = [];
 
-    constructor(private productsService: ProductsService) { }
+    constructor(
+        private productsService: ProductsService,
+        private cartService: CartService,
+        private router: Router,
+    ) { }
 
     ngOnInit(): void {
         this.loadProducts();
+    }
+
+    addProductToCart(prodcut: ProductModel): void {
+        this.cartService.addToCart(prodcut);
+    }
+
+    navigateToProduct(id: number): void {
+        this.router.navigate(['/products', id]);
     }
 
     private loadProducts(): void {