+++ changelog.md
@@ -1,6 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2022-08-08
+### Changed
+- total cost moved to cart service.
+- detection strategy.
+- display title as template var.
+
+### Added
+- the ability to add a product to the cart.
+- cart service.
+- total quantity.
+- the ability to increase ang decrease products amount.
+- highlight item in cart on hover.
+
 ## [0.0.2] - 2022-08-07
 ### Changed
 - unsibscribe.
@@ -18,3 +31,4 @@ All notable changes to this project will be documented in this file.
 
 [0.0.1]: https://github.com/illiaZakharov/Angular-34/tree/hw1
 [0.0.2]: https://github.com/illiaZakharov/Angular-34/tree/fix/task1-comments
+[0.0.3]: https://github.com/illiaZakharov/Angular-34/tree/hw2